# docs/ai-japanese-speaking-openapi.yml
openapi: 3.0.3
info:
  title: AI Japanese Speaking Service API
  version: 1.0.0
  description: |
    AI 일본어 Speaking 서비스의 1차 OpenAPI 명세입니다.
    ERD/API 기준 문서의 38개 엔드포인트와 확정된 권한·상태 규칙을 반영합니다.
    질문 수, 점수 산식, 페이징, JWT 갱신, 생성 실패 재시도 방식은 추후 확정 대상입니다.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/interviews:
    post:
      tags: [Interviews]
      summary: 면접 시작 및 첫 질문 생성
      description: 대상 국가·언어·피드백 언어는 서버가 JAPAN/JAPANESE/KOREAN으로 지정합니다.
      operationId: startInterview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewStartRequest'
      responses:
        '201':
          description: 면접 시작됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/interviews/{sessionId}/questions/{questionId}/answer:
    post:
      tags: [Interviews]
      summary: 텍스트 확정 답변 제출, 평가, Coaching 및 다음 질문 결정
      description: 답변 평가와 Coaching 후 Backend 진행 제한에 따라 Follow-up, 다음 INITIAL 또는 완료를 결정합니다.
      operationId: submitInterviewAnswer
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/QuestionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewAnswerRequest'
      responses:
        '200':
          description: 답변·평가·Coaching 저장 및 다음 질문 결정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewAnswerResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          description: 답변 평가 AI 처리 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/interviews/{sessionId}/questions/{questionId}/answer/audio:
    post:
      tags: [Interviews]
      summary: 음성 확정 답변 제출, 평가, Coaching 및 다음 질문 결정
      description: 브라우저 MediaRecorder 오디오를 STT한 뒤 답변·평가·Coaching·진행을 하나의 성공 단위로 저장합니다.
      operationId: submitInterviewAudioAnswer
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/QuestionId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: 사용자 면접 답변 녹음 파일 (webm/mp4/wav/mp3/ogg)
      responses:
        '200':
          description: STT 답변·평가·Coaching 저장 및 다음 질문 결정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewAnswerResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          description: STT 또는 답변 평가 AI 처리 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/interviews/{sessionId}/complete:
    post:
      tags: [Interviews]
      summary: 면접 정상 종료 및 전체 종합평가 생성
      description: |
        미답변 질문이 남아 있으면 완료할 수 없습니다. 종료와 종합평가 생성을 하나의 성공 단위로
        처리하므로, 종합평가 AI 생성이나 저장이 실패하면 세션도 COMPLETED로 남지 않습니다.
      operationId: completeInterview
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 종료 및 종합평가 생성 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCompletionResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          description: 종합평가 AI 처리 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/interviews/{sessionId}/abort:
    post:
      tags: [Interviews]
      summary: 면접 중도 종료
      operationId: abortInterview
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 중도 종료됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCompletionResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/interviews/{sessionId}/feedback:
    get:
      tags: [Interviews]
      summary: 면접 전체 및 답변별 피드백 조회
      operationId: getInterviewFeedback
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 면접 피드백
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewFeedback'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/organizations:
    get:
      tags: [Admin]
      summary: 기관 목록
      operationId: getOrganizations
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrganizationStatus'
      responses:
        '200':
          description: 기관 목록
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrganizationSummary'
    post:
      tags: [Admin]
      summary: 기관 생성
      description: ACTIVE 상태로 생성하며 가입 선택 목록에 포함됩니다. 기관명은 대소문자 무관 중복을 허용하지 않습니다.
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
      responses:
        '201':
          description: 생성된 기관
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationSummary'
        '409':
          description: 이미 등록된 기관명
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/organizations/{organizationId}:
    patch:
      tags: [Admin]
      summary: 기관명 수정
      operationId: updateOrganization
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
      responses:
        '200':
          description: 수정된 기관
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationSummary'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 이미 등록된 기관명
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/organizations/{organizationId}/status:
    patch:
      tags: [Admin]
      summary: 기관 상태 변경 (활성화/비활성화)
      description: >-
        하드 삭제 대신 기존 OrganizationStatus(ACTIVE/INACTIVE)를 전환하는 방식입니다.
        Organization에 소속된 User가 있어도 관계를 끊거나 User를 삭제하지 않습니다.
      operationId: changeOrganizationStatus
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrganizationStatus'
      responses:
        '200':
          $ref: '#/components/responses/OrganizationStatusChanged'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    QuestionId:
      name: questionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
...
    OrganizationStatusChanged:
      description: 기관 상태 변경 완료
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OrganizationSummary'
    BadRequest:
      description: 요청 값이 올바르지 않음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: 접근 권한 또는 소유권이 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 대상을 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 현재 상태에서 처리할 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    OrganizationStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    SessionStatus:
      type: string
      enum: [IN_PROGRESS, COMPLETED, ABORTED]
    GenerationStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED]
    Difficulty:
      type: string
      enum: [BEGINNER, INTERMEDIATE, ADVANCED]
    SubtitleMode:
      type: string
      enum: [OFF, JAPANESE, JAPANESE_KOREAN]
    QuestionKind:
      type: string
      enum: [INITIAL, FOLLOW_UP]
...
    Organization:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        status:
          $ref: '#/components/schemas/OrganizationStatus'
    OrganizationSummary:
      allOf:
        - $ref: '#/components/schemas/Organization'
        - type: object
          required: [createdAt, learnerCount]
          properties:
            learnerCount:
              type: integer
              description: 파생값. 가입 승인이 거절(REJECTED)되지 않은 LEARNER 수.
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
              nullable: true
    OrganizationUpsertRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
...
    SessionCompletionResponse:
      type: object
      required: [sessionId, status]
      properties:
        sessionId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SessionStatus'
        endedAt:
          type: string
          format: date-time
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        feedbackGenerationStatus:
          allOf:
            - $ref: '#/components/schemas/GenerationStatus'
          nullable: true
          description: abort 응답에서는 null일 수 있음
...
    InterviewStartRequest:
      type: object
      required: [jobRole, difficulty, subtitleMode]
      properties:
        jobRole:
          type: string
          maxLength: 100
        interviewType:
          type: string
          maxLength: 50
          example: TECHNICAL
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        additionalRequest:
          type: string
        subtitleMode:
          $ref: '#/components/schemas/SubtitleMode'
    InterviewQuestion:
      type: object
      required: [questionId, questionKind, questionText, sequenceNo]
      properties:
        questionId:
          type: integer
          format: int64
        questionKind:
          $ref: '#/components/schemas/QuestionKind'
        questionText:
          type: string
        sequenceNo:
          type: integer
    InterviewStartResponse:
      type: object
      required: [sessionId, status, firstQuestion]
      properties:
        sessionId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SessionStatus'
        firstQuestion:
          $ref: '#/components/schemas/InterviewQuestion'
    InterviewAnswerRequest:
      type: object
      required: [answerText]
      properties:
        answerText:
          type: string
          minLength: 1
    InterviewAnswerResponse:
      type: object
      required: [answerId, answerText, feedback, isComplete]
      properties:
        answerId:
          type: integer
          format: int64
        answerText:
          type: string
          description: STT 또는 텍스트 요청으로 확정된 사용자 답변
        feedback:
          $ref: '#/components/schemas/InterviewAnswerEvaluation'
        nextQuestion:
          allOf:
            - $ref: '#/components/schemas/InterviewQuestion'
          nullable: true
        isComplete:
          type: boolean
          description: Backend가 허용된 질문 진행을 모두 마쳤으면 true
    EvaluationCriterion:
      type: string
      enum:
        - QUESTION_RELEVANCE
        - LOGICAL_THINKING
        - SPECIFICITY
        - COMMUNICATION
        - BUSINESS_JAPANESE
        - STAR_STRUCTURE
    EvaluationScore:
      type: object
      required: [criterion, applicable]
      properties:
        criterion:
          $ref: '#/components/schemas/EvaluationCriterion'
        score:
          type: number
          format: double
          minimum: 0
          maximum: 100
          nullable: true
        feedback:
          type: string
        applicable:
          type: boolean
          description: false이면 score는 null
    InterviewAnswerEvaluation:
      type: object
      required:
        - overallScore
        - evaluationSummary
        - scores
        - strengths
        - weaknesses
        - coachingSummary
        - improvementTips
        - modelAnswer
        - rubricVersion
        - generationStatus
        - starRecommended
      properties:
        overallScore:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: applicable=true인 criterion 점수의 단순 산술 평균
        evaluationSummary:
          type: string
        scores:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationScore'
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        coachingSummary:
          type: string
          description: 기존 평가를 반복하지 않는 답변 개선 방향
        improvementTips:
          type: array
          items:
            type: string
        modelAnswer:
          type: string
          description: 사용자 답변의 사실 범위 안에서 개선한 일본어 답변. DB improved_answer에 저장됩니다.
        rubricVersion:
          type: string
        generationStatus:
          $ref: '#/components/schemas/GenerationStatus'
        starRecommended:
          type: boolean
    InterviewAnswerFeedback:
      type: object
      required: [question, answer, scores, generationStatus]
      properties:
        question:
          type: string
        answer:
          type: string
        overallScore:
          type: number
          format: double
          nullable: true
        scores:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationScore'
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        coachingSummary:
          type: string
        improvementTips:
          type: array
          items:
            type: string
        improvedAnswer:
          type: string
        rubricVersion:
          type: string
        generationStatus:
          $ref: '#/components/schemas/GenerationStatus'
    InterviewOverallFeedback:
      type: object
      properties:
        overallScore:
          type: number
          format: double
          nullable: true
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        evaluatedAnswerCount:
          type: integer
        rubricVersion:
          type: string
        generationStatus:
          $ref: '#/components/schemas/GenerationStatus'
    InterviewFeedback:
      type: object
      required: [sessionId, overall, answers]
      properties:
        sessionId:
          type: integer
          format: int64
        overall:
          $ref: '#/components/schemas/InterviewOverallFeedback'
        answers:
          type: array
          items:
            $ref: '#/components/schemas/InterviewAnswerFeedback'
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string